chore: resolve circular dependecy
