+++ ui/pages/asset/types/asset.ts
@@ -1,0 +1,33 @@
+import { Hex } from "@metamask/utils";
+import { AssetType } from "../../../../shared/constants/transaction";
+
+/** Information about a native or token asset */
+export type Asset = (
+  | {
+      type: AssetType.native;
+      /** Whether the symbol has been verified to match the chain */
+      isOriginalNativeSymbol: boolean;
+      decimals: number;
+    }
+  | {
+      type: AssetType.token;
+      /** The token's contract address */
+      address: string;
+      /** The number of decimal places to move left when displaying balances */
+      decimals: number;
+      /** An array of token list sources the asset appears in, e.g. [1inch,Sushiswap]  */
+      aggregators?: string[];
+    }
+) & {
+  /** The hexadecimal chain id */
+  chainId: Hex;
+  /** The asset's symbol, e.g. 'ETH' */
+  symbol: string;
+  /** The asset's name, e.g. 'Ethereum' */
+  name?: string;
+  /** A URL to the asset's image */
+  image: string;
+  /** True if the asset implements ERC721 */
+  isERC721?: boolean;
+  balance?: { value: string; display: string; fiat: string };
+};
